Customer Feedback Survey
This application takes basic information from the user. It has them answer a few questions but mostly wants the customer's data. So, it makes the user fill out the form and won't let them continue if there is no input. Once submitted it is sent to a database file and stored.